refactor: 결제 승인 API의 반환값을 객체에서 DTO로 수정
