+++ src/main/java/roomescape/payment/global/domain/dto/PaymentResponseDto.java
@@ -1,10 +1,10 @@
 package roomescape.payment.global.domain.dto;
 
-import roomescape.payment.global.domain.Payment;
+import roomescape.payment.global.domain.TossPayment;
 
 public record PaymentResponseDto(String orderId, int amount, String paymentKey) {
 
-    public static PaymentResponseDto of(Payment payment) {
-        return new PaymentResponseDto(payment.getOrderId(), payment.getAmount(), payment.getPaymentKey());
+    public static PaymentResponseDto from(TossPayment tossPayment) {
+        return new PaymentResponseDto(tossPayment.getOrderId(), tossPayment.getAmount(), tossPayment.getPaymentKey());
     }
 }
