+++ src/main/java/roomescape/payment/toss/service/TossPaymentService.java
@@ -5,7 +5,6 @@ import roomescape.payment.global.domain.dto.PaymentRequestDto;
 import roomescape.payment.global.domain.dto.PaymentResponseDto;
 import roomescape.payment.global.service.PaymentService;
 import roomescape.payment.toss.TossPaymentRestClient;
-import roomescape.payment.global.domain.Payment;
 
 @Service
 public class TossPaymentService implements PaymentService {
@@ -17,8 +16,6 @@ public class TossPaymentService implements PaymentService {
     }
 
     public PaymentResponseDto approve(PaymentRequestDto request) {
-        Payment payment = tossPaymentRestClient.confirmPayment(request);
-        System.out.println(payment.getPaymentType());
-        return PaymentResponseDto.of(payment);
+        return tossPaymentRestClient.confirmPayment(request);
     }
 }
