+++ src/main/resources/templates/reservation.html
@@ -11,7 +11,7 @@
     <link rel="stylesheet" href="/css/reservation.css">
     <link rel="stylesheet" href="/css/style.css">
     <link rel="stylesheet" href="/css/toss-style.css">
-    <script src="https://js.tosspayments.com/v1/payment-widget"></script>
+    <script src="https://js.tosspayments.com/v1/tossPayment-widget"></script>
 </head>
 <body>
 
@@ -87,7 +87,7 @@
 <!-- 결제창 ui -->
 <div class="wrapper w-100">
     <div class="max-w-540 w-100">
-        <div id="payment-method" class="w-100"></div>
+        <div id="tossPayment-method" class="w-100"></div>
         <div id="agreement" class="w-100"></div>
         <div class="btn-wrapper w-100">
             <button id="reserve-button" class="btn primary w-100">예약하기</button>
