+++ src/test/resources/application.properties
@@ -6,7 +6,7 @@ security.jwt.token.expire-length=3600000
 
 spring.jpa.hibernate.ddl-auto= create
 
-toss.payment.secret-key = test_gsk_docs_OaPz8L5KdmQXkzRz3y47BMw6
-toss.payment.base-url = https://api.tosspayments.com
-toss.payment.connect-timeout = 1000
-toss.payment.read-timeout = 2000
+toss.tossPayment.secret-key = test_gsk_docs_OaPz8L5KdmQXkzRz3y47BMw6
+toss.tossPayment.base-url = https://api.tosspayments.com
+toss.tossPayment.connect-timeout = 1000
+toss.tossPayment.read-timeout = 2000
